requirements.txt
appdirs==1.4.3
elasticsearch==5.2.0
MySQL-python==1.2.5
mysql-replication==0.11
packaging==16.8
PyMySQL==0.7.10
pyparsing==2.2.0
six==1.10.0
urllib3==1.20
